requirements.txt
confluent-kafka
numpy
pillow